Extract ToolingApiIdeaModelChecker
