+++ platforms/core-configuration/configuration-cache/src/integTest/groovy/org/gradle/internal/cc/impl/isolated/IsolatedProjectsToolingApiIdeaProjectIntegrationTest.groovy
@@ -22,19 +22,13 @@ import org.gradle.internal.cc.impl.actions.FetchIdeaProjectForTarget
 import org.gradle.plugins.ide.internal.tooling.idea.IsolatedIdeaModuleInternal
 import org.gradle.plugins.ide.internal.tooling.model.IsolatedGradleProjectInternal
 import org.gradle.tooling.model.idea.BasicIdeaProject
-import org.gradle.tooling.model.idea.IdeaContentRoot
-import org.gradle.tooling.model.idea.IdeaDependency
-import org.gradle.tooling.model.idea.IdeaJavaLanguageSettings
-import org.gradle.tooling.model.idea.IdeaModule
 import org.gradle.tooling.model.idea.IdeaModuleDependency
 import org.gradle.tooling.model.idea.IdeaProject
-import org.gradle.tooling.model.idea.IdeaSingleEntryLibraryDependency
 import org.gradle.tooling.provider.model.internal.PluginApplyingBuilder
 import org.gradle.util.internal.ToBeImplemented
 
-import static org.gradle.integtests.tooling.fixture.ToolingApiModelChecker.checkGradleProject
 import static org.gradle.integtests.tooling.fixture.ToolingApiModelChecker.checkModel
-import static org.gradle.integtests.tooling.fixture.ToolingApiModelChecker.checkProjectIdentifier
+import static org.gradle.internal.cc.impl.fixtures.ToolingApiIdeaModelChecker.checkIdeaProject
 
 class IsolatedProjectsToolingApiIdeaProjectIntegrationTest extends AbstractIsolatedProjectsToolingApiIntegrationTest {
 
@@ -510,71 +504,6 @@ class IsolatedProjectsToolingApiIdeaProjectIntegrationTest extends AbstractIsola
 
         // TODO:isolated check the model matches the vintage model
 //        checkIdeaProject(ideaModel, originalIdeaModel)
-    }
-
-    private static void checkIdeaProject(IdeaProject actual, IdeaProject expected) {
-        checkModel(actual, expected, [
-            { it.parent },
-            { it.name },
-            { it.description },
-            { it.jdkName },
-            { it.languageLevel.level },
-            [{ it.children }, { a, e -> checkIdeaModule(a, e) }],
-        ])
-    }
-
-    private static void checkIdeaModule(IdeaModule actualModule, IdeaModule expectedModule) {
-        checkModel(actualModule, expectedModule, [
-            { it.name },
-            [{ it.projectIdentifier }, { a, e -> checkProjectIdentifier(a, e) }],
-            [{ it.javaLanguageSettings }, { a, e -> checkLanguageSettings(a, e) }],
-            { it.jdkName },
-            [{ it.contentRoots }, { a, e -> checkContentRoot(a, e) }],
-            [{ it.gradleProject }, { a, e -> checkGradleProject(a, e) }],
-            { it.project.languageLevel.level }, // shallow check to avoid infinite recursion
-            { it.compilerOutput.inheritOutputDirs },
-            { it.compilerOutput.outputDir },
-            { it.compilerOutput.testOutputDir },
-            [{ it.dependencies }, { a, e -> checkDependency(a, e) }],
-        ])
-    }
-
-    private static void checkContentRoot(IdeaContentRoot actual, IdeaContentRoot expected) {
-        checkModel(actual, expected, [
-            { it.rootDirectory },
-            { it.excludeDirectories },
-        ])
-    }
-
-    private static void checkDependency(IdeaDependency actual, IdeaDependency expected) {
-        checkModel(actual, expected, [
-            { it.scope.scope },
-            { it.exported },
-        ])
-
-        if (expected instanceof IdeaModuleDependency) {
-            checkModel(actual, expected, [
-                { it.targetModuleName },
-            ])
-        }
-
-        if (expected instanceof IdeaSingleEntryLibraryDependency) {
-            checkModel(actual, expected, [
-                { it.file },
-                { it.source },
-                { it.javadoc },
-                { it.exported },
-            ])
-        }
-    }
-
-    private static void checkLanguageSettings(IdeaJavaLanguageSettings actual, IdeaJavaLanguageSettings expected) {
-        checkModel(actual, expected, [
-            { it.languageLevel },
-            { it.targetBytecodeVersion },
-            { it.jdk?.javaVersion },
-            { it.jdk?.javaHome },
-        ])
     }
 
     private static List<String> models(Object... models) {
